requirements.txt
discord.py~=1.7.2
pillow~=9.2.0
aiohttp>=3.7.0
